feat: break UI into multiple files
