+++ src/main.rs
@@ -3,7 +3,7 @@ use freya::prelude::{LaunchConfig, WindowConfig};
 use pg_admin::agent::{AGENT, Agent};
 use pg_admin::config::{LlmConfig, PostgresConfig, Setting};
 use pg_admin::db_client::DbClient;
-use pg_admin::ui::app;
+use pg_admin::ui::app::app;
 use tokio::sync::RwLock;
 use tracing::debug;
